backend/requirements.txt
boto3==1.34.144
channels==4.1.0
channels_redis==4.2.0
daphne==4.1.2
dj-database-url==2.2.0
django==5.0.11
django-allauth[socialaccount]==65.3.1
django-cors-headers==4.4.0
django-stubs[compatible-mypy]==5.0.2
djangorestframework==3.15.2
djangorestframework-stubs[compatible-mypy]==3.15.0
mypy==1.10.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.2
pytest-django==4.8.0
roman==5.0
